Update tkinterflow.py
I completely redid this project.  Instead of modifying tkinter's __init__.py file, I instead made a child class of each Widget (except the Menu) and then gave the child the .flow() method.  Then I reassigned the name of the parent to the child.  This only works if you first import tkinter in your project, then use the import statement for tkinterflow.  The order is important.